Move logic for discord notification from job to service
